requirements.txt
streamlit==1.52.2
pandas==2.3.3
python-dotenv==1.2.1
gspread==6.2.1
streamlit-js-eval==1.0.0